issue #689

+++ mdw-services/src/com/centurylink/mdw/services/system/SystemMetrics.java
@@ -104,7 +104,7 @@ public class SystemMetrics {
                             }
                         }
                     }
-                }, period - (LocalDateTime.now().getSecond() % period), period, TimeUnit.SECONDS);
+                }, (long) period - (LocalDateTime.now().getSecond() % period), period, TimeUnit.SECONDS);
             }
         }
     }
